refactor: add stateful downloader type
This refactor improves the flexibility of the program. Implementation
details governing *how* different artefacts are downloaded are
abstracted away by the downloader type. This makes it possible to
support a variety of link schemes while passing around a single type.

The purpose of the download module has been refined. Arguably, integrity
checking should have never been the responsibility of the download
module.

Minor changes have been made to the behaviour of the program. Integrity
checking is always performed on existing files.