refactor!: remove subcommands
Subcommands have been removed in favour of a single consistent
behaviour. This change ensures that integrity checking is always
performed.

The motivation for this change is the difficulty of implementing true
fault tolerance and recovery under the current design. The problem of
recovering from bad failures such as bad file system writes is
trivialised by always performing integrity checks. The performance
consequences of this decision are not considered to be substantial
enough to warrant a different approach. However, this decision may be
revisited later.